# api, cmd, policy: Show rule labels in `policy selectors` output