--- a/cilium/cmd/policy_selectors.go
+++ b/cilium/cmd/policy_selectors.go
@@ -12,7 +12,11 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/cilium/cilium/pkg/command"
+	k8sconst "github.com/cilium/cilium/pkg/k8s/apis/cilium.io"
+	"github.com/cilium/cilium/pkg/labels"
 )
+
+var verbosePolicySelectors bool
 
 // policyCacheGetCmd represents the policy selectors command
 var policyCacheGetCmd = &cobra.Command{
@@ -31,10 +35,22 @@ var policyCacheGetCmd = &cobra.Command{
 			sort.Slice(resp, func(i, j int) bool {
 				return resp[i].Selector < resp[j].Selector
 			})
-			fmt.Fprintf(w, "SELECTOR\tUSERS\tIDENTITIES\n")
+			fmt.Fprintf(w, "SELECTOR\tLABELS\tUSERS\tIDENTITIES\n")
+
 			for _, mapping := range resp {
+				lbls := constructLabelsArrayFromAPIType(mapping.Labels)
+
 				first := true
 				fmt.Fprintf(w, "%s", mapping.Selector)
+				if verbosePolicySelectors {
+					var lstr string
+					if len(lbls) != 0 {
+						lstr = lbls.Sort().String()
+					}
+					fmt.Fprintf(w, "\t%s", lstr)
+				} else {
+					fmt.Fprintf(w, "\t%s", getNameAndNamespaceFromLabels(lbls))
+				}
 				fmt.Fprintf(w, "\t%d", mapping.Users)
 				if len(mapping.Identities) == 0 {
 					fmt.Fprintf(w, "\t\n")
@@ -44,7 +60,7 @@ var policyCacheGetCmd = &cobra.Command{
 						fmt.Fprintf(w, "\t%d\t\n", idty)
 						first = false
 					} else {
-						fmt.Fprintf(w, "\t\t%d\t\n", idty)
+						fmt.Fprintf(w, "\t\t\t%d\t\n", idty)
 					}
 				}
 			}
@@ -54,7 +70,24 @@ var policyCacheGetCmd = &cobra.Command{
 	},
 }
 
+func getNameAndNamespaceFromLabels(lbls labels.LabelArray) string {
+	ns := lbls.Get(labels.LabelSourceK8sKeyPrefix + k8sconst.PolicyLabelNamespace)
+	if ns == "" {
+		return ""
+	}
+	return ns + "/" + lbls.Get(labels.LabelSourceK8sKeyPrefix+k8sconst.PolicyLabelName)
+}
+
+func constructLabelsArrayFromAPIType(in interface{}) labels.LabelArray {
+	lbls, ok := in.(labels.LabelArray)
+	if !ok {
+		return nil
+	}
+	return lbls
+}
+
 func init() {
+	policyCacheGetCmd.Flags().BoolVarP(&verbosePolicySelectors, "verbose", "v", false, "Show the full labels")
 	PolicyCmd.AddCommand(policyCacheGetCmd)
 	command.AddOutputOption(policyCacheGetCmd)
 }