--- a/pkg/fqdn/dnsproxy/helpers_test.go
+++ b/pkg/fqdn/dnsproxy/helpers_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/cilium/cilium/pkg/fqdn/dns"
 	"github.com/cilium/cilium/pkg/fqdn/re"
 	"github.com/cilium/cilium/pkg/identity"
+	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/policy"
 	"github.com/cilium/cilium/pkg/policy/api"
 )
@@ -212,6 +213,10 @@ type MockCachedSelector struct {
 
 func (m MockCachedSelector) GetSelections() identity.NumericIdentitySlice {
 	return nil
+}
+
+func (m MockCachedSelector) GetMetadataLabels() labels.LabelArray {
+	panic("implement me")
 }
 
 func (m MockCachedSelector) Selects(_ identity.NumericIdentity) bool {