--- a/pkg/policy/selectorcache_test.go
+++ b/pkg/policy/selectorcache_test.go
@@ -70,7 +70,7 @@ func (csu *cachedSelectionUser) AddIdentitySelector(sel api.EndpointSelector) Ca
 	csu.updateMutex.Lock()
 	defer csu.updateMutex.Unlock()
 
-	cached, added := csu.sc.AddIdentitySelector(csu, sel)
+	cached, added := csu.sc.AddIdentitySelector(csu, nil, sel)
 	csu.c.Assert(cached, Not(Equals), nil)
 
 	_, exists := csu.selections[cached]
@@ -88,7 +88,7 @@ func (csu *cachedSelectionUser) AddFQDNSelector(sel api.FQDNSelector) CachedSele
 	csu.updateMutex.Lock()
 	defer csu.updateMutex.Unlock()
 
-	cached, added := csu.sc.AddFQDNSelector(csu, sel)
+	cached, added := csu.sc.AddFQDNSelector(csu, nil, sel)
 	csu.c.Assert(cached, Not(Equals), nil)
 
 	_, exists := csu.selections[cached]
@@ -207,6 +207,9 @@ func (cs *testCachedSelector) deleteSelections(selections ...int) (deletes []ide
 
 func (cs *testCachedSelector) GetSelections() identity.NumericIdentitySlice {
 	return cs.selections
+}
+func (cs *testCachedSelector) GetMetadataLabels() labels.LabelArray {
+	return nil
 }
 func (cs *testCachedSelector) Selects(nid identity.NumericIdentity) bool {
 	for _, id := range cs.selections {
@@ -474,7 +477,7 @@ func (ds *SelectorCacheTestSuite) TestFQDNSelectorUpdates(c *C) {
 	c.Assert(len(sc.selectors), Equals, 0)
 
 	yahooSel := api.FQDNSelector{MatchName: "yahoo.com"}
-	_, added := sc.AddFQDNSelector(user1, yahooSel)
+	_, added := sc.AddFQDNSelector(user1, nil, yahooSel)
 	c.Assert(added, Equals, true)
 }